Simplify clipboard handling
