+++ packages/dashboard/src/components/editor/Editor.tsx
@@ -15,7 +15,6 @@ import { DocumentTitleEdit } from "../DocumentTitleEdit";
 import { toastController } from "@uimix/foundation/src/components/toast/Toast";
 import { LoadingErrorOverlay } from "./LoadingErrorOverlay";
 import { assertNonNull } from "../../utils/assertNonNull";
-import { DefaultClipboardHandler } from "@uimix/editor/src/state/DefaultClipboardHandler";
 
 class Connection extends TypedEmitter<{
   readyToShow(): void;
@@ -72,12 +71,6 @@ class Connection extends TypedEmitter<{
             id: documentId,
             thumbnail: url,
           });
-        },
-        getClipboard: async (type) => {
-          return await new DefaultClipboardHandler().get(type);
-        },
-        setClipboard: async (type, text) => {
-          await new DefaultClipboardHandler().set(type, text);
         },
         getCodeAssets: async () => {
           return undefined;
