+++ packages/editor/src/state/IFrameDataConnector.ts
@@ -9,7 +9,6 @@ import {
 } from "../types/IFrameRPC";
 import { throttle } from "lodash-es";
 import { ThumbnailTakerHost } from "./ThumbnailTakerHost";
-import { Clipboard } from "./Clipboard";
 import { viewOptions } from "./ViewOptions";
 
 export function vscodeParentTarget(): Target {
@@ -76,15 +75,6 @@ export class IFrameDataConnector {
     });
 
     void this.rpc.remote.ready();
-
-    // Clipboard.handler = {
-    //   get: async (type) => {
-    //     return this.rpc.remote.getClipboard(type);
-    //   },
-    //   set: async (type, text) => {
-    //     void this.rpc.remote.setClipboard(type, text);
-    //   },
-    // };
 
     void this.rpc.remote.getCodeAssets().then((assets) => {
       if (assets) {
