--- packages/editor/src/types/ClipboardHandler.ts
@@ -1,4 +1,0 @@
-export interface ClipboardHandler {
-  get(type: "text" | "image"): Promise<string | undefined>;
-  set(type: "text" | "image", textOrDataURL: string): Promise<void>;
-}
